# Solve conflict symbols of DB::ParquetBlockInputFormat and add some benchmark of optimized parquet reading