--- a/utils/local-engine/Storages/ArrowParquetBlockInputFormat.h
+++ b/utils/local-engine/Storages/ArrowParquetBlockInputFormat.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "ch_parquet/ParquetBlockInputFormat.h"
-#include "ch_parquet/ArrowColumnToCHColumn.h"
+#include "ch_parquet/OptimizedParquetBlockInputFormat.h"
+#include "ch_parquet/OptimizedArrowColumnToCHColumn.h"
 #include <Common/ChunkBuffer.h>
 #include "ch_parquet/arrow/reader.h"
 
@@ -12,12 +12,11 @@ class Table;
 
 namespace local_engine
 {
-class ArrowParquetBlockInputFormat : public DB::ParquetBlockInputFormat
+class ArrowParquetBlockInputFormat : public DB::OptimizedParquetBlockInputFormat
 {
 public:
     ArrowParquetBlockInputFormat(
         DB::ReadBuffer & in, const DB::Block & header, const DB::FormatSettings & formatSettings);
-    //virtual ~ArrowParquetBlockInputFormat();
 
 private:
     DB::Chunk generate() override;