Changelog category (leave one):
- Improvement


Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Solve conflict symbols of DB::ParquetBlockInputFormat and add some benchmark of optimized parquet reading


Benchmark result shows that parquet reading string is speeded up by 2x. 
```
---------------------------------------------------------------------------------------
Benchmark                                             Time             CPU   Iterations
---------------------------------------------------------------------------------------
BM_ParquetReadString/iterations:10                 39.9 ms         39.9 ms           10
BM_ParquetReadDate32/iterations:10                 14.3 ms         14.3 ms           10
BM_OptimizedParquetReadString/iterations:10        24.2 ms         24.2 ms           10
BM_OptimizedParquetReadDate32/iterations:200       13.4 ms         13.4 ms          200
```